+++ lib/src/observers.dart
@@ -12,34 +12,34 @@ abstract class RoutemasterObserver extends NavigatorObserver {
 
 /// Passes on navigation events to a list of [NavigatorObserver] objects.
 class _RelayingNavigatorObserver extends NavigatorObserver {
-  final List<NavigatorObserver> Function() getObservers;
+  final Iterable<NavigatorObserver> Function() getObservers;
 
   _RelayingNavigatorObserver(this.getObservers);
 
   @override
   void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didPush(route, previousRoute);
     }
   }
 
   @override
   void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didPop(route, previousRoute);
     }
   }
 
   @override
   void didRemove(Route<dynamic> route, Route<dynamic>? previousRoute) {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didRemove(route, previousRoute);
     }
   }
 
   @override
   void didReplace({Route<dynamic>? newRoute, Route<dynamic>? oldRoute}) {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didReplace(newRoute: newRoute, oldRoute: oldRoute);
     }
   }
@@ -47,14 +47,14 @@ class _RelayingNavigatorObserver extends NavigatorObserver {
   @override
   void didStartUserGesture(
       Route<dynamic> route, Route<dynamic>? previousRoute) {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didStartUserGesture(route, previousRoute);
     }
   }
 
   @override
   void didStopUserGesture() {
-    for (final observer in getObservers()) {
+    for (final observer in getObservers().toList()) {
       observer.didStopUserGesture();
     }
   }