+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.7.0
 
+* Added support for return values via `pop('Result')`
+* Added support for getting the `Route` object when pushing a new path
 * Breaking change: `Guard` now takes a `builder` method instead of a `child` property
 * Fixed an issue where subpages wouldn't push on top of a tab page
 * Fixed an issue where query string values were lost after using the back button